backend/requirements.txt
flask
flask-sqlalchemy
flask-mysqldb
flask-migrate
flask-mail
pandas
numpy
werkzeug
flask-cors